% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistribTest.R
\name{DistribTest}
\alias{DistribTest}
\title{Distribution test}
\usage{
DistribTest(x)
}
\arguments{
\item{x}{Vector of values}
}
\description{
This function runs and neatly presents distribution tests on a vector
}
\examples{
DistribTest(dataframe$vector)
}
\keyword{distribution}
\keyword{tests}
