% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalitytest.R
\name{modalitytest}
\alias{modalitytest}
\title{Modality test}
\usage{
modalitytest(x)
}
\arguments{
\item{x}{Vector of values}
}
\description{
This function runs and neatly presents modality tests on a vector
}
\examples{
modalitytest(dataframe$vector)
}
\keyword{modality}
\keyword{tests}
