% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribtest.R
\name{distribtest}
\alias{distribtest}
\title{Distribution test}
\usage{
distribtest(x)
}
\arguments{
\item{x}{Vector of values}
}
\description{
This function runs and neatly presents distribution tests on a vector
}
\examples{
distribtest(dataframe$vector)
}
\keyword{distribution}
\keyword{tests}
