% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModalityTest.R
\name{ModalityTest}
\alias{ModalityTest}
\title{Modality test}
\usage{
ModalityTest(x)
}
\arguments{
\item{x}{Vector of values}
}
\description{
This function runs and neatly presents modality tests on a vector
}
\examples{
ModalityTest(dataframe$vector)
}
\keyword{modality}
\keyword{tests}
